Android 16.0.0 release 4

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCaS5hugAKCRDorT+BmrEO
# eAMHAJ9fCttRGG2i/QYYh+XCFXYS0kDRvQCfWBnoibB6+TAp8dzuoEz+LEL8wUQ=
# =QRm4
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue 02 Dec 2025 05:49:14 IST
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

* tag 'android-16.0.0_r4': (48 commits)
  Improved usability of STATS_SOCKET_LOSS_REPORTED atom - QueueOverflow & RateLimit errors now reported only via STATS_SOCKET_LOSS_REPORTED - STATS_SOCKET_LOSS_REPORTED reported when cooldown time expired or capacity   limit is reached
  CtsStatsdHostTestCases: replace libprotobuf version
  CtsStatsdHostTestCases: statically linked host-libprotobuf-java-full
  CtsStatsdHostTestCases: removed flags usage by device side tests
  Updated statsd integration tests to remove code duplication
  Convert uses of @hide to block tag
  Fixed statsd_flags_java_lib usage in CTS helper apps
  Handle null `resultReceiver` in `onPullAtom`.
  [statsd] Configurable logging control introduced
  Fixed aconfig declaration for StatsD CTS
  Parse configs with invalid states
  Parse configs with invalid conditions
  Parse configs with invalid matchers
  ValidationTests: hold a wakelock on desktop devices
  [statsd] Added placeholder for handling atom ids in use
  [statsd] LogEventFilter refactoring
  [statsd] LogEventFilter refactoring
  [statsd] Added test to verify stats-atoms folder is created
  Fix -Wrange-loop-construct warning
  Remove Clang unused variables warnings
  ...

+++ statsd/src/anomaly/subscriber_util.h
@@ -25,7 +25,7 @@ namespace os {
 namespace statsd {
 
 void triggerSubscribers(const int64_t ruleId, int64_t metricId,
-                        const MetricDimensionKey& dimensionKey, int64_t metricValue,
+                        const MetricDimensionKey& dimensionKey, double metricValue,
                         const ConfigKey& configKey, const std::vector<Subscription>& subscriptions);
 
 }  // namespace statsd
